requirements.txt
aiogram==2.9.2
aiohttp==3.9.1
aiosignal==1.3.1
allure-pytest==2.8.18
allure-python-commons==2.8.18
annotated-types==0.6.0
attrs==23.2.0
Babel==2.14.0
certifi==2023.11.17
charset-normalizer==3.3.2
dotenv-python==0.0.1
frozenlist==1.4.1
h11==0.14.0
idna==3.6
iniconfig==2.0.0
lxml==5.1.0
multidict==6.0.4
outcome==1.3.0.post0
packaging==23.2
pluggy==1.4.0
pydantic==2.5.3
pydantic_core==2.14.6
PySocks==1.7.1
pytest==7.4.4
python-dotenv==1.0.1
requests==2.31.0
selenium==4.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tempmail-python==2.3.3
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.1.0
wsproto==1.2.0
yarl==1.9.4
